moved Pattern to editclass package
